requirements.txt
# Health Monitor - Python Dependencies
# 
# このファイルには Health Monitor アプリケーションの実行に必要な
# すべてのPython依存関係が記載されています。
#
# インストール方法:
#   pip install -r requirements.txt
#
# 仮想環境での使用を推奨:
#   python -m venv health_monitor_env
#   health_monitor_env\Scripts\activate  # Windows
#   pip install -r requirements.txt

# Core HTTP requests library for website health checks
# WebサイトのHTTPヘルスチェックに使用
requests>=2.25.0,<3.0.0

# PostgreSQL database adapter for database health checks  
# PostgreSQLデータベースの接続テストに使用
# psycopg2-binary は事前コンパイル済みバイナリを含むため、
# Windows環境でのコンパイルエラーを回避できます
psycopg2-binary>=2.8.6,<3.0.0

# Cross-platform colored terminal text for status display
# クロスプラットフォーム対応の色付きコンソール出力に使用
colorama>=0.4.4,<1.0.0

# System and process utilities for self-monitoring
# システムリソース監視とプロセス管理に使用
psutil>=5.8.0,<6.0.0

# Optional: Enhanced logging and configuration
# 将来的な機能拡張用（現在は標準ライブラリを使用）
# pyyaml>=5.4.0,<7.0.0  # YAML設定ファイルサポート用
# python-dotenv>=0.19.0,<1.0.0  # .env ファイルサポート用